+++ crates/maple_config/src/lib.rs
@@ -143,6 +143,41 @@ pub struct MarkdownPluginConfig {
     pub enable: bool,
 }
 
+#[derive(Serialize, Deserialize, Debug, Default, Eq, PartialEq)]
+#[serde(rename_all = "kebab-case", deny_unknown_fields)]
+pub enum FilePathStyle {
+    /// Display each component in path in one segment.
+    ///
+    /// crates > maple_core > src > stdio_server > plugin > ctags.rs
+    #[default]
+    OneSegmentPerComponent,
+
+    /// Display the full path in one segment.
+    ///
+    /// crates/maple_core/src/stdio_server/plugin/ctags.rs
+    FullPath,
+}
+
+/// Winbar config. (Neovim only)
+#[derive(Serialize, Deserialize, Debug, Eq, PartialEq)]
+#[serde(rename_all = "kebab-case", default, deny_unknown_fields)]
+pub struct WinbarConfig {
+    /// Whether to enable the winbar.
+    pub enable: bool,
+
+    /// Specify the display style for the file path in winbar.
+    pub file_path_style: FilePathStyle,
+}
+
+impl Default for WinbarConfig {
+    fn default() -> Self {
+        Self {
+            enable: false,
+            file_path_style: FilePathStyle::OneSegmentPerComponent,
+        }
+    }
+}
+
 /// Ctags plugin.
 #[derive(Serialize, Deserialize, Debug, Eq, PartialEq)]
 #[serde(rename_all = "kebab-case", default, deny_unknown_fields)]
@@ -361,6 +396,9 @@ pub struct Config {
     /// Log configuration.
     pub log: LogConfig,
 
+    /// Winbar configuration.
+    pub winbar: WinbarConfig,
+
     /// Matcher configuration.
     pub matcher: MatcherConfig,
 
@@ -441,6 +479,10 @@ mod tests {
                     log_file: Some("/tmp/clap.log".to_string()),
                     max_level: "trace".to_string(),
                     ..Default::default()
+                },
+                winbar: WinbarConfig {
+                    enable: false,
+                    file_path_style: FilePathStyle::OneSegmentPerComponent
                 },
                 matcher: MatcherConfig {
                     tiebreak: "score,-begin,-end,-length".to_string()